requirements.txt
anytree==2.12.1
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.6
django-admin-csvexport==2.2
django-cors-headers==4.2.0
django-import-export==3.3.6
django-modeltree==0.5
djangorestframework==3.14.0
drf-nested-routers==0.93.4
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.1
psycopg2-binary==2.9.9
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
xlrd==2.0.1
xlwt==1.3.0
